+++ docs/conf.py
@@ -302,4 +302,7 @@ texinfo_documents = [
 
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/3': None}
+intersphinx_mapping = {
+    'https://docs.python.org/3': None,
+    'https://netaddr.readthedocs.io/en/latest/': None,
+}