+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 - Feature GH-1: Add support to TAP mode of OpenVPN servers by parsing virtual
   addresses as MAC and IP both.
+- Feature GH-4: Add support to client-config-dir (ccd) and iroute.
 - Fix GH-2: **BREAK-COMPATIBILITY** Use real or virtual addresses as the key
   of client_list and routing_table, instead of using common name.